Resolve user permissions, customers, suppliers.

+++ backend-service/src/main/java/com/puff/tech/suppliermanagement/repository/SupplierRepository.java
@@ -12,10 +12,10 @@ import reactor.core.publisher.Mono;
 public interface SupplierRepository extends ReactorCrudRepository<SupplierEntity, Integer> {
 
     // Derived query: orders by createdAt descending for the given member ID
-    Flux<SupplierEntity> findByMemberIdOrderByCreatedAtDesc(Integer memberId);
+    Flux<SupplierEntity> findByMemberIdOrderByCreatedAtDesc(Long memberId);
 
     // Derived query: matches entity id and member.id
-    Mono<SupplierEntity> findByIdAndMemberId(Integer id, Integer memberId);
+    Mono<SupplierEntity> findByIdAndMemberId(Integer id, Long memberId);
 
     // Derived query: checks existence by id and member.id
     Mono<Boolean> existsByIdAndMemberId(Integer id, Integer memberId);
